backend/ethereum-api/requirements.txt
fastapi
uvicorn
web3
psycopg2-binary
redis
python-dotenv
